Implement fixes and improvements according to the requirements:
In the SideMenu component, fixed code formatting and removed unused imports.

In the Article component, created the ArticleProps interface and implemented displaying the article content.

In the News component, added passing of the selected article to the reducer using the selectArticle action.

In the Article component, added translation of the text 'Link do artykułu' to three different languages.

In the Article component, added styling in the SCSS file.

In the routes.tsx file, fixed an error in the element declaration for the '/country/:countryCode/:title' path."